+++ vitest.config.ts
@@ -3,5 +3,12 @@ import { configDefaults, defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     exclude: [...configDefaults.exclude, 'tests/browser/**'],
+    environment: 'happy-dom',
+    environmentOptions: {
+      happyDOM: {
+        url: 'http://localhost/',
+      },
+    },
+    setupFiles: ['./tests/setup.ts'],
   },
 })